[GraphQL/TransactionBlock][BUGFIX] null lamportVersion
The only situation in which the lamport version should be `null` is
when we are dealing with a system transaction that hasn't modified or
created any objects in the store.

The previous implementation always looked at created objects to find
the lamport version, but we also need to check mutated and unwrapped
objects to catch all cases.

Test Plan:

E2E tests:

```
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```