[GraphQL/TransactionBlock] Remove inputState (for now)
...until we have Object History. Currently, at best we can return
`null` for the input state, because we're viewing a transaction, so by
definition, any object input state is no longer live -- it's
historical.

Relatedly, the `IndexedObjectChange::Wrapped` was translated as having
an output state which is not possible.  The ObjectID and version in
that change represents the state of the object *before* it was
wrapped (i.e. an input state), but again, we wouldn't be able to
access it in the

Test Plan:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```