[GraphQL/MoveType] Represent using TypeTag
Previously `MoveType` was represented using its string representation.
This meant that everywhere we needed to create a `MoveType`, we had to
be careful to ensure we canonicalised the string representation first.
This PR flips things around so that the native representation relies
on `TypeTag` and we derive the canonical form of it when requested (by
the accessing the `repr` field).

This makes it harder (or impossible) to accidentally return a
non-canonical type in a query response.

Test Plan:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run -j 1 --features pg_integration
```