[TransactionEffects/API] Lamport Version
Expose lamport version from both Effects v1 and Effects v2 via the
`TransactionEffectsAPI`.  This is mainly to serve the GraphQL API
which needs to return this for both kinds of effect.

This corrects another edge case of the existing GraphQL implementation
of `lamportVersion`: If the only object created by a transaction is a
package, then the previous GraphQL implementation would incorrectly
suggest that the package's version is the transaction's lamport
version.  This is not correct, because package versioning is treated
specially.

Test Plan:

```
sui-graphql-rpc$ cargo nextest run
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- transactions
```